+++ src/pages/Portfolio/Portfolio.module.css
@@ -32,16 +32,7 @@
 .space {
   padding-bottom: 3em;
 }
-.loadingContainer {
-  width: 100%;
-  height: 35em;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.loading {
-  height: 10em;
-}
+
 @media (max-width: 391px) {
   .homeContainer {
     margin-top: 7em;
